% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_template.R
\name{add_template}
\alias{add_template}
\title{Add Rmd template}
\usage{
add_template(
  template = NULL,
  save_as = template,
  overwrite = FALSE,
  open = FALSE
)
}
\arguments{
\item{template}{Various templates to choose from.}

\item{save_as}{Name of output file.}

\item{overwrite}{Logical. Whether to overwrite existing template file.}

\item{open}{Logical. Whether to open file after creation.}
}
\value{
Create Rmd file(s) template(s) and return its (their) path
}
\description{
Choose from various templates to be added to your project.
}
\examples{
# For classical use in your package
\dontrun{
# add only the dev_history file in an existing package
add_template()
}
}
