% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadLibs.R
\name{downloadLibs}
\alias{downloadLibs}
\title{Download Libraries}
\usage{
downloadLibs()
}
\value{
a working computer (hopefully...)
}
\description{
\code{downloadLibs} downloads libraries that are used for quanting
}
\examples{
\dontrun{downloadLibs()}
}
