% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_rprofile_template.R
\name{fct_add_rprofile_template}
\alias{fct_add_rprofile_template}
\title{Add Rprofile template}
\usage{
fct_add_rprofile_template(scope = c("user", "project"))
}
\arguments{
\item{scope}{Scope of the .Rprofile file to edit.}
}
\value{
.rprofile template content copied to clipboard.
}
\description{
Add Rprofile template
}
\examples{
if (FALSE) {
  fct_add_rprofile_template()
}
}
\seealso{
\code{\link[usethis]{edit}}
}
