% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixLibPath.R
\name{fixLibPath}
\alias{fixLibPath}
\title{Fix Library Path}
\usage{
fixLibPath()
}
\value{
awesomeness
}
\description{
\code{fixLibPath} sets up your R installation to use the Q drive
library. You will need to restart R for the changes to take effect.
}
\examples{
\dontrun{fixLibPath()}
}
