% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupR-package.R
\docType{package}
\name{setupR-package}
\alias{setupR}
\alias{setupR-package}
\title{setupR: Templates and Workflows for Data Science Projects}
\description{
Use templates to make common data science workflows easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fabiandistler/setupR}
  \item Report bugs at \url{https://github.com/fabiandistler/setupR/issues}
}

}
\author{
\strong{Maintainer}: Fabian Distler \email{fabian.distler@gmx.net} (\href{https://orcid.org/0000-0002-1438-0724}{ORCID})

}
\keyword{internal}
