% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_template.R
\name{add_template}
\alias{add_template}
\title{Add Rmd template}
\usage{
add_template(
  template = c("dev_history"),
  pkg = ".",
  dev_dir = "dev",
  flat_name = NULL,
  overwrite = FALSE,
  open = TRUE
)
}
\arguments{
\item{template}{Name of the template to use. See details.}

\item{pkg}{Path where to save file}

\item{dev_dir}{Name of directory for development Rmarkdown files. Default to "dev".}

\item{flat_name}{Name of the file to write in dev.}

\item{overwrite}{Whether to overwrite existing Rmd template file with same name}

\item{open}{Logical. Whether to open file after creation}
}
\value{
Create flat Rmd file(s) template(s) and return its (their) path
}
\description{
Add Rmd template
}
\details{
This function was adapted from the awesome \code{{fusen}} package.

Choose \code{template} among the different templates available:
\itemize{
\item "dev_history": Template with functions commonly used during package development.
This does not contain chunks to write your own functions.
}
}
\examples{
# For classical use in your package
\dontrun{
# add only the dev_history file in an existing package
add_template()
}
}
